tweetビューのアバターにtooltipのメニューが出るようにする #25
